trying to bind

+++ benchmarking/socket/forward/proxy_udp.c
@@ -72,7 +72,7 @@ int main(int argc, char *argv[]){
                   alarm(duration);
             }
             send_adr.sin_addr.s_addr = inet_addr("192.168.1.2");
-            n = sendto(sockfd,buffer,size,0,(struct sockaddr*) &send_adr,len);
+            n = sendto(sockfd,buffer,size,0,(struct sockaddr*) &send_adr,(ssize_t )len);
             if(n<=0){
                   if(n==0)
                         printf("sended zero bytes\n");
